+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Added new code validation for *NotImplementedError* to raise a warning in `XSOAR-linter`.
 * Added validation for support types in the pack metadata file.
 * Fixed an issue with running **validate** on master branch where the changed files weren't compared to previous commit when using the *-g* flag.
+* Fixed an issue where the `XSOAR-linter` ran *NotImplementedError* validation on scripts.
 
 # 1.2.12
 * Bandit now reports also on medium severity issues.